Add files via upload
In this we have connected to the database and created singIn panel, Login panel, and a dashboard. The user signs in and their data gets stored in database and when the user tries to login there is checking between the input and data in the database. Then opens a dashboard with the latest tax news and another tab where a user can save owns profile information.